Rename folders to make them deployable to ESP32.
